--- src/main/java/gift/controller/GlbalExceptionHandler.java
@@ -1,35 +1,0 @@
-package gift.controller;
-
-import gift.exception.MemberAuthorizationException;
-import jakarta.persistence.EntityNotFoundException;
-import java.util.ArrayList;
-import java.util.List;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlbalExceptionHandler {
-
-    @ExceptionHandler(value = MemberAuthorizationException.class)
-    public ResponseEntity<String> handleMemberException(Exception e) {
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.UNAUTHORIZED);
-    }
-
-    @ExceptionHandler(EntityNotFoundException.class)
-    public ResponseEntity<?> handleEntityNotFoundException(EntityNotFoundException e) {
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(value = MethodArgumentNotValidException.class)
-    public ResponseEntity<List<String>> handleNameException(MethodArgumentNotValidException e) {
-        List<String> errorMessages = new ArrayList<>();
-        e.getBindingResult().getAllErrors().forEach(error -> {
-            String errorMessage = error.getDefaultMessage();
-            errorMessages.add(errorMessage);
-        });
-        return new ResponseEntity<>(errorMessages, HttpStatus.BAD_REQUEST);
-    }
-}
