--- src/main/java/gift/domain/Wish.java
@@ -1,34 +1,0 @@
-package gift.domain;
-
-public class Wish {
-
-    private Long id;
-    private Long memberId;
-    private Long productId;
-
-    public Wish() {
-    }
-
-    public Wish(Long memberId, Long productId) {
-        this(null, memberId, productId);
-    }
-
-    public Wish(Long id, Long memberId, Long productId) {
-        this.id = id;
-        this.memberId = memberId;
-        this.productId = productId;
-    }
-
-    public Long getId() {
-        return this.id;
-    }
-
-    public Long getMemberId() {
-        return memberId;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-
-}
