--- src/main/java/gift/dto/JwtResponse.java
@@ -1,14 +1,0 @@
-package gift.dto;
-
-public class JwtResponse {
-
-    private String token;
-
-    public JwtResponse(String token) {
-        this.token = token;
-    }
-
-    public String getToken() {
-        return token;
-    }
-}
