--- src/main/java/gift/dto/MemberRequest.java
@@ -1,25 +1,0 @@
-package gift.dto;
-
-public class MemberRequest {
-
-    String email;
-    String password;
-
-    public MemberRequest() {
-    }
-
-    public MemberRequest(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-
-}
