--- src/main/java/gift/dto/ProductIdRequest.java
@@ -1,17 +1,0 @@
-package gift.dto;
-
-public class ProductIdRequest {
-
-    private Long productId;
-
-    public ProductIdRequest() {
-    }
-
-    public ProductIdRequest(Long productId) {
-        this.productId = productId;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-}
