+++ src/main/java/gift/member/dto/MemberRequest.java
@@ -1,0 +1,25 @@
+package gift.member.dto;
+
+public class MemberRequest {
+
+    String email;
+    String password;
+
+    public MemberRequest() {
+    }
+
+    public MemberRequest(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+
+}
