+++ src/main/java/gift/product/controller/ProductController.java
@@ -1,0 +1,63 @@
+package gift.product.controller;
+
+import gift.product.dto.ProductRequest;
+import gift.product.service.ProductService;
+import gift.product.domain.Product;
+import jakarta.validation.Valid;
+import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+
+@RestController
+@RequestMapping("/api/products")
+public class ProductController {
+
+    private final ProductService productService;
+
+    public ProductController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping()
+    public ResponseEntity<List<Product>> readAll(
+        @RequestParam(defaultValue = "0") int pageNo,
+        @RequestParam(defaultValue = "10") int pageSize
+    ) {
+        return new ResponseEntity<>(productService.getAllProducts(pageNo, pageSize), HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Product> read(@PathVariable("id") Long id) {
+        return new ResponseEntity<>(productService.getProduct(id), HttpStatus.OK);
+    }
+
+    @PostMapping()
+    public ResponseEntity<Product> create(@RequestBody @Valid ProductRequest productRequest) {
+        Product product = productService.addProduct(productRequest);
+        return new ResponseEntity<>(product, HttpStatus.CREATED);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Product> update(@PathVariable("id") Long id,
+        @RequestBody @Valid ProductRequest productRequest) {
+        Product product = productService.updateProduct(id, productRequest);
+
+        return new ResponseEntity<>(product, HttpStatus.OK);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Product> delete(@PathVariable("id") Long id) {
+        productService.deleteProduct(id);
+        return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
+    }
+}
