+++ src/main/java/gift/product/domain/Product.java
@@ -1,0 +1,37 @@
+package gift.product.domain;
+
+public class Product {
+
+    private Long id;
+    private String name;
+    private int price;
+    private String imageUrl;
+
+    public Product() {
+    }
+
+    public Product(Long id, String name, int price, String imageUrl) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.imageUrl = imageUrl;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getPrice() {
+        return this.price;
+    }
+
+    public String getImageUrl() {
+        return this.imageUrl;
+    }
+
+
+}
