+++ src/main/java/gift/product/dto/ProductResponse.java
@@ -1,0 +1,5 @@
+package gift.product.dto;
+
+public class ProductResponse {
+
+}
