+++ src/main/java/gift/product/repository/ProductRepository.java
@@ -1,0 +1,12 @@
+package gift.product.repository;
+
+import gift.product.entity.ProductEntity;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
+    Page<ProductEntity> findAll(Pageable pageable);
+}
