--- src/main/java/gift/repository/MemberRepository.java
@@ -1,13 +1,0 @@
-package gift.repository;
-
-import gift.entity.MemberEntity;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
-
-    Optional<MemberEntity> findByEmail(String email);
-
-}
