+++ src/main/java/gift/util/dto/JwtResponse.java
@@ -1,0 +1,14 @@
+package gift.util.dto;
+
+public class JwtResponse {
+
+    private String token;
+
+    public JwtResponse(String token) {
+        this.token = token;
+    }
+
+    public String getToken() {
+        return token;
+    }
+}
