+++ src/main/java/gift/wishlist/dto/ProductIdRequest.java
@@ -1,0 +1,17 @@
+package gift.wishlist.dto;
+
+public class ProductIdRequest {
+
+    private Long productId;
+
+    public ProductIdRequest() {
+    }
+
+    public ProductIdRequest(Long productId) {
+        this.productId = productId;
+    }
+
+    public Long getProductId() {
+        return productId;
+    }
+}
