+++ src/main/java/gift/wishlist/dto/WishRequest.java
@@ -1,0 +1,24 @@
+package gift.wishlist.dto;
+
+public class WishRequest {
+
+    private Long memberId;
+
+    private Long productId;
+
+    public WishRequest() {
+    }
+
+    public WishRequest(Long memberId, Long productId) {
+        this.memberId = memberId;
+        this.productId = productId;
+    }
+
+    public Long getMemberId() {
+        return memberId;
+    }
+
+    public Long getProductId() {
+        return productId;
+    }
+}
