+++ src/main/java/gift/wishlist/dto/WishResponse.java
@@ -1,0 +1,5 @@
+package gift.wishlist.dto;
+
+public class WishResponse {
+
+}
