+++ src/test/java/gift/controller/RegisterTest.java
@@ -2,7 +2,7 @@ package gift.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gift.dto.MemberRequest;
+import gift.member.dto.MemberRequest;
 import java.net.URI;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
