+++ src/test/java/gift/repository/MemberRepositoryTest.java
@@ -3,8 +3,9 @@ package gift.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
-import gift.dto.MemberRequest;
-import gift.entity.MemberEntity;
+import gift.member.dto.MemberRequest;
+import gift.member.entity.MemberEntity;
+import gift.member.repository.MemberRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
