+++ src/test/java/gift/repository/ProductRepositoryTest.java
@@ -3,8 +3,9 @@ package gift.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
-import gift.dto.ProductRequest;
-import gift.entity.ProductEntity;
+import gift.product.repository.ProductRepository;
+import gift.product.dto.ProductRequest;
+import gift.product.entity.ProductEntity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
