+++ src/test/java/gift/service/WishServiceTest.java
@@ -7,18 +7,17 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import gift.domain.Member;
-import gift.domain.Product;
-import gift.domain.Wish;
-import gift.dto.MemberRequest;
-import gift.dto.WishRequest;
-import gift.entity.MemberEntity;
-import gift.entity.ProductEntity;
-import gift.entity.WishEntity;
-import gift.repository.MemberRepository;
-import gift.repository.ProductRepository;
-import gift.repository.WishRepository;
-import java.lang.reflect.Array;
+import gift.member.domain.Member;
+import gift.product.domain.Product;
+import gift.wishlist.domain.Wish;
+import gift.wishlist.dto.WishRequest;
+import gift.member.entity.MemberEntity;
+import gift.product.entity.ProductEntity;
+import gift.wishlist.entity.WishEntity;
+import gift.member.repository.MemberRepository;
+import gift.product.repository.ProductRepository;
+import gift.wishlist.repository.WishRepository;
+import gift.wishlist.service.WishService;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
